supplier dao, service and controller added

+++ src/main/java/com/inventory/deva_inventory/dao/SupplierRepository.java
@@ -14,6 +14,6 @@ import org.springframework.stereotype.Repository;
  * @author best
  */
 @Repository
-public interface SupplierRepository extends JpaRepository<Supplier, Long>{
+public interface SupplierRepository extends JpaRepository<Supplier, Integer>{
     
 }
